Make `Cache-Control` headers configurable
The default SaltAPI CORS settings don't allow the client to set a
`Cache-Control` header, which this application does unconditionally.

Either SaltAPI should be changed (because setting the `no-cache` header
value makes a lot of sense), or we should take a risk and disable it at
the application side.

This commit makes this configurable.